refactor(vps): update vCPU, RAM, and disk fields to allow decimal values
Changed the step attribute for NumberField components in the VpsEditSheet to "any" for vCPU, RAM, and disk inputs. Updated the vpsSchema to accept real numbers for vCPU, diskGb, and bandwidthTb. Adjusted database schema to store vCPU, diskGb, and bandwidthTb as REAL instead of INTEGER for better precision.

+++ apps/web/src/lib/schemas.ts
@@ -50,7 +50,7 @@ export const vpsSchema = z.object({
   country: z.string().optional().default(''),
   city: z.string().optional().default(''),
   datacenter: z.string().optional().default(''),
-  vcpu: z.coerce.number().int().min(0).default(1),
+  vcpu: z.coerce.number().min(0).default(1),
   ramGb: z.coerce.number().min(0).default(1),
   diskGb: z.coerce.number().min(0).default(10),
   status: vpsStatusSchema.default('active'),
